requirements.txt
llama-index-core
llama-index-llms-openai
llama-index-readers-file
llama-index-readers-web
unstructured[pdf,docx,pptx]
paddleocr
pillow
trafilatura
cohere
pypdf
python-pptx
sentence-transformers
rank_bm25
llama-index-embeddings-openai
llama-index-vector-stores-faiss
llama-index
faiss-cpu
fitz
frontend
nest-asyncio
python-dotenv